+++ debug/main.ml
@@ -78,7 +78,7 @@ let debug_resolver () =
   print_endline (To_sexp.to_sexp_string timere);
   let search_start_dt =
     Time.Date_time'.make ~year:2000 ~month:`Jan ~day:1 ~hour:10 ~minute:0
-      ~second:0 ~tz
+      ~second:0 ~tz ()
     |> CCOpt.get_exn
   in
   let search_start =
@@ -87,7 +87,7 @@ let debug_resolver () =
   in
   let search_end_exc_dt =
     Time.Date_time'.make ~year:2050 ~month:`Jan ~day:1 ~hour:0 ~minute:0
-      ~second:0 ~tz
+      ~second:0 ~tz ()
     |> CCOpt.get_exn
   in
   let search_end_exc =
@@ -101,10 +101,10 @@ let debug_resolver () =
           timere;
           after
             (Date_time'.make_exn ~tz ~year:2000 ~month:`Jan ~day:1 ~hour:0
-               ~minute:0 ~second:0);
+               ~minute:0 ~second:0 ());
           before
             (Date_time'.make_exn ~tz ~year:2050 ~month:`Jan ~day:1 ~hour:0
-               ~minute:0 ~second:0);
+               ~minute:0 ~second:0 ());
           (* intervals [ (search_start, search_end_exc)
            *           ] *)
         ])
@@ -151,10 +151,10 @@ let debug_example () =
               ymdhms 2020 Jun 16 10 0 0)) (points (pick dhms 17 12 0 0)))";
            after
              (Date_time'.make_exn ~tz ~year:2000 ~month:`Jan ~day:1 ~hour:0
-                ~minute:0 ~second:0);
+                ~minute:0 ~second:0 ());
            before
              (Date_time'.make_exn ~tz ~year:2050 ~month:`Jan ~day:1 ~hour:0
-                ~minute:0 ~second:0);
+                ~minute:0 ~second:0 ());
          ])
   in
   match Resolver.resolve timere with